requirements.txt
beautifulsoup4
biopython
requests
